+++ dotCMS/src/main/java/com/dotcms/cdi/business/LogTimeInterceptor.java
@@ -1,4 +1,4 @@
-package com.dotcms.business.cdi;
+package com.dotcms.cdi.business;
 
 import com.dotcms.business.interceptor.LogTimeHandler;
 import com.dotcms.util.LogTime;
